ml-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
tensorflow==2.20.0
pillow==10.1.0
numpy==2.0.0
python-dotenv==1.0.0
